Feat: ShoppingCart컴포넌트 코드를 cart페이지로 이동
- 기존 cart 페이지 코드를 OriginalCartPage 컴포넌트로 분리
