+++ src/app/(home)/cart/page.tsx
@@ -3,16 +3,23 @@
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
-import Button from "@/components/BeforeEditByKi/Button/Button";
+import BottomButton from "@/components/BottomButton/BottomButton";
+import Button from "@/components/Button/Button";
+import ShoppingCartCard from "@/components/Card/ShoppingCartCard";
+import PriceSummaryCard from "@/components/PriceCheckCard/PriceSummaryCard";
+import TopNavigator from "@/components/TopNavigator/TopNavigator";
 
 import { AccessoryItem } from "@/store/Items/accessoryStore";
 import { CabinetItem } from "@/store/Items/cabinetStore";
 import { DoorItem } from "@/store/Items/doorStore";
 import { FinishItem } from "@/store/Items/finishStore";
 import { HardwareItem } from "@/store/Items/hardwareStore";
-import { DeliverTime } from "@/utils/CheckDeliveryTime";
 
-import CartItemDetail from "./_components/CartItemDetail";
+const DOOR_TYPE_KR_MAP: Record<string, string> = {
+  normal: "일반문",
+  flap: "플랩문",
+  drawer: "서랍",
+};
 
 const CATEGORY_MAP: Record<string, string> = {
   door: "문짝",
@@ -24,11 +31,9 @@ const CATEGORY_MAP: Record<string, string> = {
 
 type OrderItem = DoorItem | FinishItem | CabinetItem | AccessoryItem | HardwareItem | null;
 
-export default function CartPage() {
+function ShoppingCart() {
   const router = useRouter();
 
-  const [deliveryMessage, setDeliveryMessage] = useState("");
-  const [deliveryMessageColor, setDeliveryMessageColor] = useState("text-black");
   const [cartGroups, setCartGroups] = useState<Record<string, OrderItem[]>>({});
   const [hasItems, setHasItems] = useState(false);
 
@@ -40,41 +45,13 @@ export default function CartPage() {
     const grouped: Record<string, OrderItem[]> = {};
     cartItems.forEach(item => {
       if (!item) return;
-      if (!grouped[item.category]) {
-        grouped[item.category] = [];
-      }
+      if (!grouped[item.category]) grouped[item.category] = [];
       grouped[item.category].push(item);
     });
+
     setCartGroups(grouped);
     setHasItems(cartItems.length > 0);
-
-    const fetchDeliveryTime = async () => {
-      if (address1 !== "주소 없음") {
-        const { expectedArrivalMinutes } = await DeliverTime(address1);
-        const cutoff = 18 * 60;
-        const hours = Math.floor(expectedArrivalMinutes / 60)
-          .toString()
-          .padStart(2, "0");
-        const minutes = (expectedArrivalMinutes % 60).toString().padStart(2, "0");
-        if (expectedArrivalMinutes <= cutoff) {
-          setDeliveryMessage(`당일배송 가능 ${hours}:${minutes}`);
-          setDeliveryMessageColor("bg-[#cbdcfb] text-[#215cff]");
-        } else {
-          setDeliveryMessage(`내일 배송되는 주소에요`);
-          setDeliveryMessageColor("bg-gray-500 text-[#bf6a02]");
-        }
-      }
-    };
-    fetchDeliveryTime();
   }, []);
-
-  const handleOrder = () => {
-    router.push("/cart/checkorder");
-  };
-
-  const handleAddProduct = () => {
-    router.push("/");
-  };
 
   const getTotalPrice = () => {
     return Object.values(cartGroups)
@@ -102,82 +79,183 @@ export default function CartPage() {
     });
   };
 
+  const handleOrder = () => {
+    router.push("/cart/checkorder");
+  };
+
+  const handleAddProduct = () => {
+    router.push("/");
+  };
+
   if (!hasItems) {
     return (
-      <div className="flex h-[80vh] flex-col items-center justify-center p-5">
-        <h1 className="mb-4 text-xl font-bold">장바구니</h1>
-        <p className="mb-6 text-center text-lg font-semibold">
-          상품이 하나도 없어요.
-          <br />
-          상품을 담아볼까요?
-        </p>
-        <Button className="w-full bg-gray-200 text-black" onClick={handleAddProduct}>
-          상품 추가
-        </Button>
+      <div className="flex min-h-screen flex-col">
+        <TopNavigator title="장바구니" />
+        <div className="flex flex-1 flex-col items-center justify-center px-5">
+          <img src="/icons/paper.svg" alt="빈 용지 아이콘" className="mb-3" />
+          <p className="mb-5 text-center text-[17px] font-500 text-gray-500">장바구니가 비었어요</p>
+          <div>
+            <Button text="상품추가" type="Brand" onClick={handleAddProduct}></Button>
+          </div>
+        </div>
       </div>
     );
   }
 
+  const getTotalItemCount = () => {
+    return Object.values(cartGroups)
+      .flat()
+      .reduce((sum, item) => sum + (item?.count ?? 0), 0);
+  };
+  const sanitizedCartGroups = Object.fromEntries(
+    Object.entries(cartGroups).map(([key, items]) => [key, items.filter(Boolean)]),
+  );
   return (
-    <div className="flex flex-col p-5 pb-20">
-      <h1 className="mb-4 text-center text-xl font-bold">장바구니</h1>
+    <div className="flex min-h-screen flex-col">
+      <TopNavigator title="장바구니" />
+      <div className="flex-1 overflow-y-auto pb-[150px]">
+        <div className="p-5">
+          <div className="pb-3 text-xl font-600">상품 {getTotalItemCount()}개</div>
+          {Object.entries(cartGroups).map(([category, items], groupIdx) => (
+            <div key={category + groupIdx} className="mb-4 flex flex-col gap-3">
+              {items.map((item, i) => {
+                if (!item) return null;
 
-      <div className={`mb-4 p-2 text-center font-medium ${deliveryMessageColor}`}>
-        {deliveryMessage}
+                const key = `${category}-${i}`;
+                const commonProps = {
+                  trashable: true,
+                  onIncrease: () => handleCountChange(category, i, (item.count ?? 0) + 1),
+                  onDecrease: () => handleCountChange(category, i, (item.count ?? 0) - 1),
+                };
+
+                if (category === "door") {
+                  const doorItem = item as DoorItem;
+
+                  return (
+                    <ShoppingCartCard
+                      key={key}
+                      title={DOOR_TYPE_KR_MAP[doorItem.slug ?? "normal"]}
+                      color={doorItem.color ?? ""}
+                      width={String(doorItem.width ?? "")}
+                      height={String(doorItem.height ?? "")}
+                      hingeCount={doorItem.hinge?.hingeCount ?? 0}
+                      hingeDirection={doorItem.hinge?.hingePosition === "left" ? "좌경" : "우경"}
+                      boring={`상 ${doorItem.hinge?.topHinge} 중 ${doorItem.hinge?.middleHinge} 하 ${doorItem.hinge?.bottomHinge}`}
+                      quantity={doorItem.count ?? 0}
+                      trashable={true}
+                      onIncrease={() => handleCountChange(category, i, (doorItem.count ?? 0) + 1)}
+                      onDecrease={() => handleCountChange(category, i, (doorItem.count ?? 0) - 1)}
+                    />
+                  );
+                }
+
+                if (category === "finish") {
+                  const finishItem = item as FinishItem;
+                  return (
+                    <ShoppingCartCard
+                      key={key}
+                      title={"마감재"}
+                      color={finishItem.color ?? ""}
+                      width={""}
+                      height={String(finishItem.height ?? "")}
+                      hingeCount={0}
+                      hingeDirection={"없음"}
+                      boring={"-"}
+                      quantity={finishItem.count ?? 0}
+                      {...commonProps}
+                    />
+                  );
+                }
+
+                if (category === "cabinet") {
+                  const cabinetItem = item as CabinetItem;
+                  return (
+                    <ShoppingCartCard
+                      key={key}
+                      title={cabinetItem.slug ?? "부분장"}
+                      color={cabinetItem.color ?? ""}
+                      width={String(cabinetItem.width ?? "")}
+                      height={String(cabinetItem.height ?? "")}
+                      hingeCount={0}
+                      hingeDirection={"없음"}
+                      boring={"-"}
+                      quantity={cabinetItem.count ?? 0}
+                      {...commonProps}
+                    />
+                  );
+                }
+
+                if (category === "accessory") {
+                  const accessoryItem = item as AccessoryItem;
+                  return (
+                    <ShoppingCartCard
+                      key={key}
+                      title={accessoryItem.slug ?? "부속품"}
+                      color={"-"}
+                      width={"-"}
+                      height={"-"}
+                      hingeCount={0}
+                      hingeDirection={"없음"}
+                      boring={"-"}
+                      quantity={accessoryItem.count ?? 0}
+                      {...commonProps}
+                    />
+                  );
+                }
+
+                if (category === "hardware") {
+                  const hardwareItem = item as HardwareItem;
+                  return (
+                    <ShoppingCartCard
+                      key={key}
+                      title={hardwareItem.slug ?? "하드웨어"}
+                      color={"-"}
+                      width={"-"}
+                      height={"-"}
+                      hingeCount={0}
+                      hingeDirection={"없음"}
+                      boring={"-"}
+                      quantity={hardwareItem.count ?? 0}
+                      {...commonProps}
+                    />
+                  );
+                }
+
+                return null;
+              })}
+            </div>
+          ))}
+          <BottomButton
+            type="1button"
+            button1Text="상품 추가"
+            button1Type="BrandInverse"
+            className="w-full pt-0"
+            onButton1Click={handleAddProduct}
+          />
+        </div>
+        <div className="px-5">
+          <PriceSummaryCard
+            cartGroups={sanitizedCartGroups}
+            getTotalPrice={getTotalPrice}
+            categoryMap={CATEGORY_MAP}
+          />
+        </div>
       </div>
 
-      {Object.entries(cartGroups).map(([category, items], groupIdx) => (
-        <div key={category + groupIdx} className="mb-4 flex flex-col gap-3">
-          {items.map((item, i) => {
-            if (!item) return null;
-            return (
-              <CartItemDetail
-                key={`${category}-${i}`}
-                item={item}
-                onCountChange={newCount => handleCountChange(category, i, newCount)}
-              />
-            );
-          })}
-        </div>
-      ))}
-
-      <Button selected={true} className="mb-3" onClick={handleAddProduct}>
-        상품 추가
-      </Button>
-
-      <div className="my-6">
-        <p className="mb-2 text-lg font-semibold">결제금액을 확인해주세요</p>
-
-        <div className="w-full bg-gray-300 p-4">
-          <div className="mb-2 flex justify-between">
-            <span className="font-bold">총 금액</span>
-            <span className="font-bold">{getTotalPrice().toLocaleString()}원</span>
-          </div>
-          {Object.entries(cartGroups).map(([category, items]) => {
-            const categoryTotal = items.reduce((sum, item) => {
-              if (!item) return sum;
-              return sum + (item.price ?? 0) * (item.count ?? 1);
-            }, 0);
-
-            return (
-              <div key={category} className="mb-1 flex justify-between">
-                <span>{CATEGORY_MAP[category] || category}</span>
-                <span>{categoryTotal.toLocaleString()}원</span>
-              </div>
-            );
-          })}
-
-          <hr className="my-2 border-black" />
-
-          <div className="flex justify-between">
-            <span className="font-bold">결제예정금액</span>
-            <span className="font-bold">{getTotalPrice().toLocaleString()}원</span>
-          </div>
-        </div>
+      <div className="w-full max-w-[500px] bg-white">
+        <BottomButton
+          type="textcombo+button"
+          textComboText={{
+            title: `${getTotalPrice().toLocaleString()}원`,
+            subtitle: "주문금액",
+          }}
+          button1Text="다음"
+          button1Type="Brand"
+          onButton1Click={handleOrder}
+        />
       </div>
-      <Button selected={true} onClick={handleOrder}>
-        다음
-      </Button>
     </div>
   );
 }
+
+export default ShoppingCart;
